src/user_editor/backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic
python-jose[cryptography]
passlib[argon2]
argon2-cffi